src/requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
et-xmlfile==1.1.0
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.3
openai==0.19.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.0
six==1.16.0
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
# this needs to be updated to the latest version of youtube-dl using github master
git+https://github.com/ytdl-org/youtube-dl.git@master#egg=youtube-dl
whisper.ai==1.0.0.1
dropbox==11.36.0
azure-cognitiveservices-speech==1.25.0
oauthlib==3.2.2 
requests-oauthlib==1.3.1 
tweepy==4.12.1
pyrebase4==4.6.0
audioread==3.0.0
google-api-python-client==2.77.0
google-auth-oauthlib==1.0.0
flask==2.0.2
gspread>=5.0.0
replicate==0.5.0
ShopifyAPI==12.2.0
praw==7.7.0
